Move deconz function to util.py (#136414)
